+++ src/charts/src/common/get-color.ts
@@ -1,0 +1,6 @@
+export const getColor = (
+  categoryMap: Map<string, string>,
+  category: string
+) => {
+  return categoryMap.get(category) ?? "#4f46e5";
+};