+++ src/charts/tailwind.config.js
@@ -6,7 +6,6 @@ export const content = [
   "./components/**/*.{ts,tsx}",
   "./app/**/*.{ts,tsx}",
   "./src/**/*.{ts,tsx}",
-  "../components/**/*.{ts,tsx}",
 ];
 
 export const theme = {