+++ src/components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # maison-components-core
 
+## 3.1.0
+
+### Minor Changes
+
+- optional datepicker trigger
+
 ## 3.0.0
 
 ### Major Changes