+++ src/form/CHANGELOG.md
@@ -1,5 +1,12 @@
 # maison-components-form
 
+## 2.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - maison-components-core@3.1.0
+
 ## 2.2.0
 
 ### Minor Changes